refactor(tasks): move Go2W joystick owner [skip ci]
