+++ src/unilab/envs/locomotion/go2w/__init__.py
@@ -1,1 +1,1 @@
-from .joystick import Go2WJoystickCfg, Go2WJoystickEnv
+"""Legacy Go2W shared base pending migration to :mod:`unilab.tasks`."""
