+++ src/unilab/tasks/__init__.py
@@ -10,8 +10,7 @@ and deterministic throughout the migration.
 __unilab_registry_modules__ = (
     "unilab.tasks.locomotion.go1",
     "unilab.tasks.locomotion.go2",
-    "unilab.envs.locomotion.go2w",
-    "unilab.tasks.locomotion.go2w.rough",
+    "unilab.tasks.locomotion.go2w",
     "unilab.envs.locomotion.g1",
     "unilab.envs.locomotion.go2_arm",
     "unilab.tasks.locomotion.a2",
