+++ src/unilab/tasks/locomotion/go2w/__init__.py
@@ -1,3 +1,9 @@
+from .joystick import Go2WJoystickCfg, Go2WJoystickEnv
 from .rough import Go2WJoystickRoughCfg, Go2WJoystickRoughEnv
 
-__all__ = ["Go2WJoystickRoughCfg", "Go2WJoystickRoughEnv"]
+__all__ = [
+    "Go2WJoystickCfg",
+    "Go2WJoystickEnv",
+    "Go2WJoystickRoughCfg",
+    "Go2WJoystickRoughEnv",
+]
