+++ src/unilab/tasks/locomotion/go2w/rough.py
@@ -32,7 +32,7 @@ from unilab.envs.locomotion.common.terrain_spawn import (
     TerrainSpawnManager,
 )
 from unilab.envs.locomotion.go2w.base import NUM_GO2W_ACTIONS, NUM_LEG_ACTIONS
-from unilab.envs.locomotion.go2w.joystick import (
+from unilab.tasks.locomotion.go2w.joystick import (
     Go2WJoystickCfg,
     Go2WJoystickDomainRandomizationProvider,
     Go2WJoystickEnv,
